requirements.txt
ipykernel
python-dotenv
cassio
datasets
langchain
openai
tiktoken
PyPDF2
langchain_community
streamlit
langchain_openai
astrapy